Added Validation to LLM configuration
Created a new 'llm_functions.py' scrypt which holds functions for utilise all LLM's, checking configuration is set correctly, updating usability flag, and updating stored llm's configuration.

I've move the updating_llm_configuration main logic to the new file as it makes more senese logically and cleans up the setting spage code a bit.

I've also modified the save buttons to run the validiation check when pressed, and if it passes (gets a response from the llm) then the configuration is saved. The store buttons colour is based on the "usable" trait for each LLM in the config file, and changes to green when the validation is complete.